requirements.txt
numpy
pandas
scikit-learn
Pillow
tensorflow
os